Lab 1 - submission in progress
Solved tasks 1, 2A and 2B
In progress 2C and 3

Tests 5, 6 and 7 are test cases for 2C and 3. I had previously written solutions for them in the code, but I removed them after I kept getting errors that were unexplainable to me. I tried implementing 2C by adding a counter for silences from the user in the IS, and updating that throughout the dialog. This updating and incrementing was quite confusing, and I though it be best to hand in the assignment and discuss the VG parts later.